Test Refactoring Removing API to get test runtime provider from TestEngine
