+++ src/Microsoft.TestPlatform.Common/Interfaces/Engine/ClientProtocol/ITestEngine.cs
@@ -39,13 +39,5 @@ namespace Microsoft.VisualStudio.TestPlatform.ObjectModel.Engine
         /// </summary>
         /// <returns>ITestExtensionManager object that helps with extensibility</returns>
         ITestExtensionManager GetExtensionManager();
-
-        /// <summary>
-        /// Fetches the Test Host manager for this engine. This manager would provide extensibility
-        /// features that this engine supports.
-        /// </summary>
-        /// <param name="runSettingsXml">runSettingsXml information which contains info like Architecture, Framework for the test run.</param>
-        /// <returns>Launcher for the test host process</returns>
-        ITestRuntimeProvider GetDefaultTestHostManager(string runSettingsXml);
     }
 }
